+++ src/modules/check/src/lib.rs
@@ -1,3 +1,4 @@
 #[cfg(feature = "cli")]
 pub mod cli;
-pub mod core;
+pub mod logic;
+pub mod types;