+++ src/modules/generator/src/core/mod.rs
@@ -1,3 +1,1 @@
-pub mod interfaces;
 pub mod service;
-pub mod template_engine;