+++ src/modules/templatespecification/src/core/mod.rs
@@ -2,4 +2,5 @@ pub mod file_list;
 pub mod interfaces;
 pub mod service;
 pub mod template_configuration;
+pub mod template_engine;
 pub mod template_specification;